All features in.
Create an application that allows users to submit pictures of cakes. All the cakes may be viewed, rated from 1-5 stars, and commented upon. When a cake image is clicked, have a nested details component show the image, the baker of the cake, the average rating, and all the different ratings and comments.